implementado metodos do append  . iniciado correção de bugs no appendat

+++ ApplicationService/Contracts/AppendAt.cs
@@ -6,9 +6,9 @@ namespace ApplicationService.Contracts
     public class AppendAt : OperationContract
     {
 
-        public long _position { get; private set; }
+        public int _position { get; private set; }
         public string _text { get; private set; }
-        public AppendAt(long position, string text, List<FileIdentity> files, string dir)
+        public AppendAt(int position, string text, List<FileIdentity> files, string dir)
             : base(dir, files)
         {
             _position= position;
